requirements.txt
brotlicffi==1.0.9.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.60.1
ftfy==6.3.1
gmpy2==2.2.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.7
# mkl_fft==2.1.2
# mkl_random==1.3.0
# mkl-service==2.5.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opencv-python==4.6.0.66
packaging==25.0
pillow==11.1.0
pip==25.2
pycparser==2.23
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
# torch==2.1.2
# torchaudio==2.1.2
# torchvision==0.16.2
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
wheel==0.45.1
win_inet_pton==1.1.0
yacs==0.1.8

# app
httpx==0.27.2
aiofiles==24.1.0
fastapi==0.111.1
pyheif==0.8.0
pandas==2.0.3